### 2024-01-21 #6252 @ckdckd145
@arfon

Thanks for your kind explanation. I really sympathize with the busy you mentioned. I'll wait until a reviewer is assigned.

I have a few quick questions before the review begins. If you have a moment, I'd really appreciate it if you could answer them.

1. is it okay to make changes and updates in submitted software before the review starts?

2. is it okay to make changes to paper.bib as well as paper.md if there are any changes?

Thanks!
