+++ article/paper.bib
@@ -14,5 +14,5 @@
   pages={171--185},
   year={2003},
   publisher={Taylor & Francis},
-  doi={https://doi.org/10.1080/0307507032000058064}
+  doi={10.1080/0307507032000058064}
 }
