### 2024-01-21 #6252 @arfon
> is it okay to make changes and updates in submitted software before the review starts?

Yes, this is an ideal time to be making changes.

> is it okay to make changes to paper.bib as well as paper.md if there are any changes?

Yes.
