### 2024-01-31 968d04e
bug fixed in selector parameter
if selector parameters were changed and run in the same line, i find that that dataframe were overlap-filtered. To fix this, add codes for init selector parameters were provided.
