+++ statmanager/manager.py
@@ -29,7 +29,9 @@ class Stat_Manager:
     def __init__(self, dataframe: pd.DataFrame, id: str = None, language: str = 'kor'):
         self.df = dataframe
         self.df_original = dataframe
+        self.df_analysis = None
         self.filtered_df = None
+        self.conditions = None
         self.result = None
         self.selector = None
         
@@ -76,7 +78,7 @@ class Stat_Manager:
         """
         Please check the documentation : https://cslee145.notion.site/statmanager-kr-Documentation-c9d0886f29ea461d9d0f44449a145f8a?pvs=4
         
-        """
+        """        
         method = method.lower()
         posthoc_method = posthoc_method.lower()
 
@@ -147,9 +149,11 @@ class Stat_Manager:
         
         if selector == None:
             self.selector = None
-            df = self.df_original
+            self.df_analysis = self.df_original
+            #df = self.df_original
             self.filtered_df = None          
-            conditions = None
+            self.conditions = None
+            #conditions = None
             self.conditions_notification_texts = None
         
         else:
@@ -157,9 +161,15 @@ class Stat_Manager:
             if type(selector) != dict:
                 raise TypeError(error_message_for_selector_type(doclink = self.link)[self.language_set])
             
-            df = self.df
+            #always init the selector and related variables for repeated running
+            self.selector = None
+            self.df_analysis = self.df_original
+            self.filtered_df = None          
+            self.conditions = None
+            self.conditions_notification_texts = None
+            
             self.selector = selector
-            self.filtered_df = None 
+            
             conditions  = []
             conditions_notification = []
             
@@ -167,40 +177,41 @@ class Stat_Manager:
                 if isinstance(value, dict):  
                     for op, op_value in value.items():
                         if op == '<=':
-                            conditions.append(df[key] <= op_value)
+                            conditions.append(self.df_analysis[key] <= op_value)
                             conditions_notification.append(f"{key} <= {op_value}")
                         elif op == '>=':
-                            conditions.append(df[key] >= op_value)
+                            conditions.append(self.df_analysis[key] >= op_value)
                             conditions_notification.append(f"{key} >= {op_value}")
                         elif op == '<':
-                            conditions.append(df[key] < op_value)
+                            conditions.append(self.df_analysis[key] < op_value)
                             conditions_notification.append(f"{key} < {op_value}")
                         elif op == '>':
-                            conditions.append(df[key] > op_value)
+                            conditions.append(self.df_analysis[key] > op_value)
                             conditions_notification.append(f"{key} > {op_value}")
                         elif op == '=' or op == '==':
-                            conditions.append(df[key] == op_value)
+                            conditions.append(self.df_analysis[key] == op_value)
                             conditions_notification.append(f"{key} == {op_value}")
                         elif op == '!=':
-                            conditions.append(df[key] != op_value)
+                            conditions.append(self.df_analysis[key] != op_value)
                             conditions_notification.append(f"{key} != {op_value}")
 
                 else: 
-                    conditions.append(df[key] == value)
+                    conditions.append(self.df_analysis[key] == value)
                     conditions_notification.append(f"{key} == {value}")
             
             combined_condition = conditions[0]
             
-            if len(conditions) == 1: #selector 하나만 붙인 경우에는 그냥 진행 
+            if len(conditions) == 1: # when the selector is only one 
                 pass
             
-            else: # selector가 2개 이상인 경우 
+            else: # when the selector are more than one
                 for cond in conditions[1:]:
[…]
