### 2024-02-01 a5e4037
Improving readability of results printed in ttest_ind_yuen, mannwhitneyu, brunner. like ttest_ind
