+++ statmanager/between_group_functions.py
@@ -74,6 +74,7 @@ def ttest_ind_yuen(df: pd.DataFrame, vars: list or str, group_vars : str, lang_s
     
     result_for_save = []
     trim_not_working = False
+    result_df = pd.DataFrame(columns = ['dependent variable', 't-value', 'degree of freedom', 'p-value', '95% CI']).set_index('dependent variable')
     
     dv = vars[0] if isinstance(vars, list) else vars
     
@@ -93,6 +94,14 @@ def ttest_ind_yuen(df: pd.DataFrame, vars: list or str, group_vars : str, lang_s
     p = result_object.pvalue
     dof = result_object.df
     ci = result_object.confidence_interval() 
+    
+    result_df.loc[dv] = [s, dof, p, f"[{ci.low:.3f}, {ci.high:.3f}]"]
+    
+    for _ in result_df.columns:
+        if _ != '95% CI':
+            result_df[_] = result_df[_].astype(float).round(3)
+        else:
+            continue
     
     # trimmed_data
     
@@ -131,11 +140,11 @@ def ttest_ind_yuen(df: pd.DataFrame, vars: list or str, group_vars : str, lang_s
     
     reporting_one = compare_btwgroup_result_reporting_one(dv, group_vars, group_names)[lang_set]
     
-    reporting_two = ttest_ind_result_reporting_two (s, p, dof, ci, cohen_d)[lang_set]
+    # reporting_two = ttest_ind_result_reporting_two (s, p, dof, ci, cohen_d)[lang_set]
     result_for_save.append(notation)
     result_for_save.append(reporting_one)
     result_for_save.append(describe_df)
-    result_for_save.append(reporting_two)
+    result_for_save.append(result_df)
     
     print(testname)
     
@@ -150,10 +159,9 @@ def ttest_ind_yuen(df: pd.DataFrame, vars: list or str, group_vars : str, lang_s
                 
     return result_for_save    
     
-    
-    
 def mannwhitneyu(df: pd.DataFrame, vars: list or str, group_vars : str, lang_set, testname, posthoc = None, posthoc_method = None):
     result_for_save = []
+    result_df = pd.DataFrame(columns = ['dependent variable', 'U-value', 'Z-value', 'p-value', 'Rank-biserial Correlation']).set_index('dependent variable')
     
     dv = vars[0] if isinstance(vars, list) else vars
     group_vars = group_vars[0] if isinstance(group_vars, list) else group_vars
@@ -178,12 +186,17 @@ def mannwhitneyu(df: pd.DataFrame, vars: list or str, group_vars : str, lang_set
     z = (s - n1 * n2 / 2) / ((n1 * n2 * (n1 + n2 + 1)) / 12)**0.5
     rank_biserial_correlation = 1 - (2 * s) / (n1 * n2)
     
+    result_df.loc[dv, : ] = [s, z, p, rank_biserial_correlation]
+    
+    for _ in result_df.columns:
+        result_df[_] = result_df[_].astype(float).round(3)
+    
     reporting_one = compare_btwgroup_result_reporting_one(dv, group_vars, group_names)[lang_set]
-    reporting_two = compare_btwgroup_result_reporting_two (s, p, z, rank_biserial_correlation)[lang_set]
+    # reporting_two = compare_btwgroup_result_reporting_two (s, p, z, rank_biserial_correlation)[lang_set]
     
     result_for_save.append(reporting_one)
     result_for_save.append(describe_df)
-    result_for_save.append(reporting_two)
+    result_for_save.append(result_df)
     
     print(testname)
     for n in result_for_save:
@@ -199,6 +212,7 @@ def mannwhitneyu(df: pd.DataFrame, vars: list or str, group_vars : str, lang_set
 
[…]
