### 2024-02-02 4d463c1
Improving readibility of results printed in ttest_rel, wilcoxon, friedman
