### 2024-02-15 817d23f
deleting unnecessary _ in dummy coded columns in regression methods
