### 2024-02-15 e7cbaec
improvement in oneway ANCOVA : now auto dummy coded functions were activated when nominal variables were provided as covariates
