### 2024-02-15 0e98be9
add docstring for Stat_Manager class and .progress()
