### 2024-02-16 6c25781
same improvement in repeated measures ANCOVA (covariates will automatically dummy coded)
