+++ statmanager/ancova_functions.py
@@ -19,7 +19,7 @@ def oneway_ancova(df:pd.DataFrame, vars: list, lang_set, testname, posthoc, post
     
     iv = group_vars # str
     
-    covar_df = pd.get_dummies(df[covars], drop_first=True, dtype='float', prefix = 'dummy_')
+    covar_df = pd.get_dummies(df[covars], drop_first=True, dtype='float', prefix = 'dummy')
     
     new_covars = []
     for covar in covars:
@@ -125,41 +125,57 @@ def oneway_ancova(df:pd.DataFrame, vars: list, lang_set, testname, posthoc, post
 def rm_ancova(df:pd.DataFrame, vars: list, group_vars, lang_set, testname, posthoc, posthoc_method):
     
     result_for_save = []
-    
+
     index_col = df.index.name
     repeated_vars = vars [:-1]
+    vars_for_customjoin = vars [:-1]
     covars = vars[-1]
-    vars_for_melting  = [index_col] + covars
 
-    decribe_vars = ['value'] + covars
+    covar_df = pd.get_dummies(df[covars], drop_first = True, dtype='float', prefix = 'dummy')
 
+    new_covars = []
+    for covar in covars:
+        if covar in covar_df.columns:
+            covar_df = covar_df.drop(columns=covar)
+            new_covars.append(covar)
+    new_covars += covar_df.columns.to_list()
+
+    df = df.merge(covar_df, left_index=True, right_index = True, how = 'left')
+
+    vars_for_customjoin.append(new_covars)
+    print(vars_for_customjoin)
+
+    vars_for_melting  = [index_col] + new_covars
+
+    decribe_vars = ['value'] + new_covars
 
     melted_df = df.reset_index().melt(id_vars = index_col, value_vars = repeated_vars, var_name = 'time')
     melted_df = melted_df.merge(df.reset_index()[vars_for_melting], on = index_col, how = 'left').set_index(index_col)
-    
-    formula_for_olsmodel = custom_join_for_ancova(vars = vars, method = 'rm_ancova')
+
+    formula_for_olsmodel = custom_join_for_ancova(vars = vars_for_customjoin, method = 'rm_ancova')
     olsmodel = ols(formula_for_olsmodel, data = melted_df).fit()
-    
+
     describe_df = melted_df.groupby('time')[decribe_vars].agg(['count', 'mean', 'median', 'std']).rename(
                     columns = {
                         'count' : 'n',
                         'std' : 'sd'
                     }).T.round(2)
     describe_df.columns.name = None
-    
+
     anova_table = anova_lm(olsmodel, typ=3)
     anova_table.rename(columns = {'PR(>F)' : 'p-value'}, inplace=True)
     anova_table['partial_eta_squared'] = anova_table['sum_sq'] / (anova_table['sum_sq'] + anova_table['sum_sq'].loc['Residual'])
     anova_table = anova_table.round(3)
-    
+
     raw_coef_table = pd.DataFrame(olsmodel.summary().tables[1].data, columns = ['index','coef', 'std err', 't', 'p', '0.025', '0.975'])[1:].set_index('index')
 
     pair_coef_table = raw_coef_table.loc[['Intercept']]
-    covar_coef_table = raw_coef_table.loc[covars]
-    drop_col_for_coef_table = ['Intercept'] + covars
-    
+    covar_coef_table = raw_coef_table.loc[new_covars]
+    drop_col_for_coef_table = ['Intercept'] + new_covars
+
+
     for n in range( len (melted_df.time.unique() ) ):
-        formula_for_coef = custom_join_for_ancova(vars = vars, method = 'rm_ancova', purpose = 'coef', keys = n)
[…]
