### 2024-02-16 2a53c4a
add and improvement of docstrings in manager.py
