+++ statmanager/manager.py
@@ -26,21 +26,45 @@ VERSION = __version__
 LINK = LINK_DOC
 
 class Stat_Manager: 
-    def __init__(self, dataframe: pd.DataFrame, id: str = None, language: str = 'kor'):
-        """ This is a mandatory class that must be created in order to use statmanager-kr.   
-        If you haven't read the officical documentation, it is recommended to check this: [Documentation](https://cslee145.notion.site/60cbfcbc90614fe990e02ab8340630cc?v=4991650ae5ce4427a215d1043802f5c0&pvs=4)
+    """
+    Summary
+    ----
+    This is a mandatory class that must be created in order to use statmanager-kr.   
+    If you haven't read the officical documentation, it is recommended to check this: [Documentation](https://cslee145.notion.site/60cbfcbc90614fe990e02ab8340630cc?v=4991650ae5ce4427a215d1043802f5c0&pvs=4)
+    
+    Parameters:
+    ----
+    dataframe (pandas.DataFrame):   
+        - The pandas.DataFrame to which statistical analysis will conducted.   
+        - The DataFrame must follow a wide-range shape.   
+        
+    id (str, optional): Defaults to None.   
+        - If the DataFrame has no index columns, specify the index.   
+        
+    language (str, optional): Defaults to 'kor'.   
+        - Language settings. Supported Languages are English ('eng') and Korean ('kor').
+            
+            
+    Functions:
+    ----
+    `.progress()`
+        - Main method to conduct statistical analysis
+        - For more details, see the docstrings of .progress() or [Documentation](https://cslee145.notion.site/60cbfcbc90614fe990e02ab8340630cc?v=4991650ae5ce4427a215d1043802f5c0&pvs=4)
+    
+    `.howtouse()`
+        - Method to see how to use the statmanager-kr
+        - For more details, see the docstrings of .howtouse()
+        
+    `.set_language()`
+        - Method to change language setting
+        - For more details, see the docstrings of .set_language()
 
-        Parameters:
-            dataframe (pandas.DataFrame):   
-                - The pandas.DataFrame to which statistical analysis will conducted.   
-                - The DataFrame must follow a wide-range shape.   
-                
-            id (str, optional): Defaults to None.   
-                - If the DataFrame has no index columns, specify the index.   
-                
-            language (str, optional): Defaults to 'kor'.   
-                - Language settings. Supported Languages are English ('eng') and Korean ('kor').   
-        """
+    `.change_dataframe()`
+        - Method to change dataframe
+        - For more details, see the docstrings of .change_dataframe()
+    """
+    def __init__(self, dataframe: pd.DataFrame, id: str = None, language: str = 'kor'):
+
         self.df = dataframe
         self.df_original = dataframe
         self.df_analysis = None
@@ -90,6 +114,8 @@ class Stat_Manager:
         
     def progress(self, method: str, vars: list, group_vars: str = None, posthoc: bool = False, posthoc_method: str = 'bonf', selector: dict = None):
         """
+        Summary
+        ----
         Method to apply the statistical anaylsis based on the key argument provided to method parameter.   
         The results of the analysis will printed in format of str and pandas.DataFrame.   
         Also, an object of StatmanagerResult class will returned.   
@@ -98,87 +124,90 @@ class Stat_Manager:
         If you haven't read the officical documentation, it is recommended to check this: [Documentation](https://cslee145.notion.site/60cbfcbc90614fe990e02ab8340630cc?v=4991650ae5ce4427a215d1043802f5c0&pvs=4)
            
         Parameters:
-            method (str): Key value of statistical analysis   
[…]
