### 2024-02-19 8efd628
add docstring in StatmanagerResult class and its methods
