+++ README.md
@@ -1,5 +1,5 @@
  
-[![pypi](https://badge.fury.io/py/statmanager-kr.svg?kill_cache?kill=cache)](https://badge.fury.io/py/statmanager-kr)
+[![pypi](https://badge.fury.io/py/statmanager-kr.svg)](https://badge.fury.io/py/statmanager-kr)
 [![license](https://img.shields.io/badge/license-MIT-green.svg)](https://github.com/ckdckd145/statmanager-kr/blob/main/LICENSE)
 ![logo](./doc/logo.png)
 
