### 2024-02-27 #6252 @ckdckd145
Hi @sappelhoff  . Thank you for taking the time and effort to check my project and provide feedback. Following your suggestion, I have looked at pingouin and recognize that there are some very similarities in purpose and approach. Pingouin is definitely a great statistical package, but I think statmanager-kr has some unique advantages that it can offer to other users.

Therefore, I will add a comparison paragraph to the paper as per your suggestion and also introduce some more "user-friendly" features of statmanager-kr. After I finish working on the paper, I will regenerate the PDF via editorial bot and inform you (if you don't mind)!

Once again, thank you for your feedback.
Have a great day!
