+++ article/paper.bib
@@ -16,3 +16,34 @@
   publisher={Taylor & Francis},
   doi={10.1080/0307507032000058064}
 }
+@article{Virtanen:2020,
+  title={SciPy 1.0: fundamental algorithms for scientific computing in Python.}
+  author={Virtanen, Pauli and Gommers, Ralf and Oliphant, Travis E and Haberland, Matt and Reddy, Tyler and Cournapeau, David and Burovski, Evgeni and Peterson, Pearu and Weckesser, Warren and Bright, Jonathan and others},
+  journal={Nature methods},
+  volume={17},
+  number={3},
+  pages={261--272},
+  year={2020},
+  publisher={Nature Publishing Group},
+  doi={10.1038/s41592-019-0686-2}
+}
+@article{seabold:2010,
+  title={Statsmodels: Econometric and statistical modeling with python},
+  author={Seabold, Skipper and Perktold, Josef},
+  booktitle={Proceedings of the 9th Python in Science Conference},
+  volume={57},
+  number={61},
+  pages={10--25080},
+  year={2010},
+  organization={Austin, TX}
+}
+@article{vallat2018,
+  title={Pingouin: statistics in Python.},
+  author={Vallat, Raphael},
+  journal={Journal of Open Source Software},
+  volume={3},
+  number={31},
+  pages={1026},
+  year={2018},
+  doi={10.21105/joss.01026}
+}
