+++ article/paper.bib
@@ -17,7 +17,7 @@
   doi={10.1080/0307507032000058064}
 }
 @article{Virtanen:2020,
-  title={SciPy 1.0: fundamental algorithms for scientific computing in Python.}
+  title={SciPy 1.0: fundamental algorithms for scientific computing in Python.},
   author={Virtanen, Pauli and Gommers, Ralf and Oliphant, Travis E and Haberland, Matt and Reddy, Tyler and Cournapeau, David and Burovski, Evgeni and Peterson, Pearu and Weckesser, Warren and Bright, Jonathan and others},
   journal={Nature methods},
   volume={17},
@@ -30,12 +30,13 @@
 @article{seabold:2010,
   title={Statsmodels: Econometric and statistical modeling with python},
   author={Seabold, Skipper and Perktold, Josef},
-  booktitle={Proceedings of the 9th Python in Science Conference},
+  journal={Proceedings of the 9th Python in Science Conference},
   volume={57},
   number={61},
   pages={10--25080},
   year={2010},
-  organization={Austin, TX}
+  organization={Austin, TX},
+  doi={10.25080/Majora-92bf1922-011}
 }
 @article{vallat2018,
   title={Pingouin: statistics in Python.},
