### 2024-02-28 #6252 @sappelhoff
Great, thanks. I have another comment that might help you:

Unlike `pingouin`, which is a software with an [extensive test suite](https://github.com/raphaelvallat/pingouin/tree/master/tests), your software only features very little (if any?) testing. Am I just not seeing these tests, or are they not implemented? I think you should invest time in building a thorough test suite for your package. This will not only increase the reliability of your software, but it's also a requirement for submissions to JOSS.
