### 2024-02-28 #6252 @ckdckd145
@sappelhoff

Thanks again for your comment. I agree with you that a package should provide enough functionality for its purpose, and I also know that this is a requirement of JOSS. As you said, I am not sure if you have seen all the features of statmanager-kr, so I share the feature table below.

| Key                | Analysis                                       |
| ------------------ | ---------------------------------------------- |
| `kstest`           | Kolmogorov-Smirnov Test                        |
| `shapiro`          | Shapiro-Wilks Test                             |
| `z_normal`         | Z-skeweness & z-kurtosis test                  |
| `levene`           | Levene Test                                    |
| `fmax`             | Fmax Test                                      |
| `chi2_contingency` | Chi-squared Test                               |
| `fisher`           | Fisher's Exact Test                            |
| `pearsonr`         | Pearson's r                                    |
| `spearmanr`        | Spearman's rho                                 |
| `kendallt`         | Kendall's tau                                  |
| `ttest_ind`        | Independent Samples T-test                     |
| `ttest_rel`        | Dependent Samples T-test                       |
| `ttest_ind_trim`   | Yuen's Two Samples T-test                      |
| `ttest_ind_welch`  | Welch's Two Samples T-test                     |
| `mannwhitneyu`     | Mann-Whitney U Test                            |
| `brunner`          | Brunner-Munzel Test                            |
| `wilcoxon`         | Wilcoxon-Signed Rank Test                      |
| `bootstrap`        | Boostrap Percentile Method                     |
| `f_oneway`         | One-way ANOVA                                  |
| `f_oneway_rm`      | One-way Repeated Measures ANOVA                |
| `kruskal`          | Kruskal-Wallis Test                            |
| `friedman`         | Friedman Test                                  |
| `f_nway`           | N-way ANOVA                                    |
| `f_nway_rm`        | N-way Mixed Repeated Measures ANOVA            |
| `linearr`          | Linear Regression                              |
| `hier_linearr`     | Hierarchical Linear Regression                 |
| `logisticr`        | Logistic Regression                            |
| `oneway_ancova`    | One-way ANCOVA                                 |
| `rm_ancova`        | One-way Repeated Measures ANCOVA               |
| `cronbach`         | Calculating Cronbach's Alpha                   |
| `bonf`             | Bonferroni Correction                          |
| `tukey`            | Tukey HSD                                      |

If this is the case for you, then statmanager-en may not be sufficient for the needs of JOSS. However, I've been using statmanager-kr to create reports and other tasks in my current research, including a government-funded research project. Therefore, actually I think it has a value in social science research. But, if it is lacking in functionality, I will ask for a review in the future. So it would be great if you could look at it again.

Thanks again for your feedback!
