### 2024-02-28 #6252 @sappelhoff
sorry for the misunderstanding @ckdckd145 --> I was not talking about "statistical tests" or the features of your software, but about ["unit tests"](https://en.wikipedia.org/wiki/Unit_testing) that ensure that your software is working as expected. One of the ways `pingouin` does this, for example, is through comparisons with tests implemented in the `R` language. See also: https://docs.pytest.org/en/8.0.x/
