### 2024-03-01 #6252 @ckdckd145
@sappelhoff

Ah, now that I understand your comment, I agree with you. To be honest, I hadn't thought of that.

Like Pingouin, I'll add the information to the documentation. It will take some time, but fortunately statmanager-kr is on the waiting list, and I'll get back to you for further review and advice once it's done.

Thank you very much.
Have a great day.
