### 2024-03-04 #6252 @ckdckd145
@sappelhoff

In response to your feedback, I've added comparison documentation where possible.

You can view them one by one by clicking on the analysis columns in the table at [HERE](https://cslee145.notion.site/Documentation-74a610c12881402d96dc5d1654f97433#be93db7f4159419fa73eb324d6567793). There are still some analyses that are incomplete, I'll continue to add them.

Thanks again !
