### 2024-03-11 18c4810
badge update and error type fix
