### 2024-03-11 4bf7df2
add standardized regression coef. in linearr & hier_linearr
