### 2024-03-18 b673048
remove round in regression analysis table. To prevent some case that table printing 0.000 (with too small value)
