### 2024-03-19 #6252 @catstats
Thank you for considering me for the review of the submission to JOSS. I am happy to contribute to the review process. Please kindly let me know the next steps and any specific timelines or details I should be aware of.

Best,
castats
