### 2024-04-16 #6252 @behinger
I was travelling, thanks for waiting for my response. I can review it, but earliest starting next week.

As a disclaimer: I havent used much of pythons stats ecosystem in the last years. But if I read the JOSS Critera correct, it is not a problem if other software already exists doing a similar job.
