### 2024-04-18 #6642 @editorialbot
[REVIEW]: Statmanager-kr: A user-friendly statistical package for python in pandas
**Submitting author:** @ckdckd145 (<a href="http://orcid.org/0000-0002-8825-5571">Changseok Lee</a>)
**Repository:** https://github.com/ckdckd145/statmanager-kr
**Branch with paper.md** (empty if default branch):
**Version:** v1.8.1.14
