### 2024-04-30 ckdckd145/statmanager-kr#12 @behinger
JOSS Reviewer Issue
Hi!
nice software you have there. I will use this issue to keep track of problems I found on the way to use it.

- [x] the differences to pengouin are mentioned in the paper, but I would also add them to the readme under "related software" with a brief discussion of strengths of each
- [x] a compatible python version should be provided. I'm running my tests with python 3.12 current version
- [x] clarification: "with df that have a structure of WIDE-RANGE. " => is WIDE-RANGE the same as a tidy dataframe in R?
- [x] while the errormessage is generally good (it tells me what I need to do), I don't understand why I need to specify an index for the DF.
- [x]  `Search sm.howtouse("fgiure") for the function to draw pictures and graphs! ` => typo in figure
- [x] in the `howtouse()` output it would be nice to get an example how to actual use a method, e.g. something like `sm.progress(method = 'ttest_ind', vars = 'age', group_vars = 'sex').figure()` - (not for every method, just one general example to get an idea how to use it)
- [x] if I define an id="species", I cannot use it anymore as grouping variable
- [x] if I specify more than 2 groups in ttest_ind, I get a very cryptic error `AxisError: `axis` must be an integer, a tuple of integers, or `None`.`
- [x] typo: `Indenpendent` in howtouse
- [x] if I run `ttest_rel` I get a `KeyError: 's'` if I use the wrong call syntax (which is impossible to find out from the REPL/jupyter notebook, you have to look into the actual manual).
- [x] automated testing: While in the docs in notion there are some printed comparisons against e.g. scipy, it is not clear whether these are run automatically (I didnt find the code for the documentation), or whether the author checked each for equivalence
- [x] in a spot-check I saw that the cohen's d calculation was implemented by the author, but I cannot confirm that there exists a test for these functions.
- [x] The JOSS Part " Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support" is not completed as of now (or I missed it)

I will continue later, but the unittesting is really important, especially in a stats-package

Linking back to (openjournals/joss-reviews/issues/6642)
