### 2024-04-30 ckdckd145/statmanager-kr#13 @catstats
JOSS review issue2
I've been exploring your package and appreciate its capabilities. However, I encountered a couple of issues that I would like to bring to your attention:

1. The use of index_col="name" as suggested in the Quickstart guide does not work with the provided testdata/testdf.csv. It functions correctly when I switch to another variable, such as "id".

2. The bar plot generated by the command sm.progress(method='ttest_ind', vars='age', group_vars='sex').figure() appears misaligned. Is there a specific version of the plotting library that I should be using to ensure proper alignment?
