### 2024-04-30 #6642 @catstats
[https://github.com/ckdckd145/statmanager-kr/issues/13](url)
