### 2024-05-02 #6642 @ckdckd145
@catstats @behinger

Hi ! Thanks for reviewing my article and source code of the statmanager-kr.
Also thank you very much for overall positive feedbacks (I hope I read correctly)

I briefly checked all comments and now planning to revise all related parts soon.
I will change the codes, and whenever I commit, I'll mention it in the related issue in my repository.
But, I think this to take a while, so I'd appreciate your patience. (Thanks!)

To be honest, I'm not a developer by trade, and I'm new to the Joss review process, so I'm not sure if this is the right way to do it. Please let me know if this is something I should do differently.
Thank you so much. :)

ref @teonbrooks
