### 2024-05-14 2288fc1
README REVISION - Typo error in Quickstart section (id = 'name' -> id = 'id') - Add related software section (Especially for Pingouin)
