+++ README.md
@@ -5,14 +5,14 @@
 
 
 # Statmanager-kr
-### Open-source statistical package for Python based on the Pandas.    
+### Open-source Statistical Package for Python based on the Pandas.    
 ### Python과 Pandas 사용자를 위한 오픈소스 통계 패키지
  
 #
 
-Especially for researchers, data scientists, psychologist, students, and anyone who interested in conducting hypothesis testing. The statmanager-kr aims to organize packages that are "convenient to use", "uncompliated to use", and "convenient to see results". The end goal of statmanager-kr is to be a <b>simple and useful package</b> that <b>can be used </b>by people <b>who don't know much about Python and Pandas</b>.  
+Statmanager-kr is developed for researchers, data scientists, psychologist, studends, and anyone who need statistical analysis to validate their hypothesis. The statmanager-kr aims to organize packages that are "convenient to use", "uncompliated to use", and "convenient to see results". The end goal of statmanager-kr is to be a <b>simple and useful package</b> that <b>can be used </b>by people <b>who don't know much about Python and Pandas</b>.  
 
-Pandas를 사용하며, 가설 검증에 대해 관심을 갖는 연구원, 데이터분석가, 심리학자, 학생 등을 위합니다. statmanager-kr은 사용하기 쉽고, 사용이 복잡하지 않으며, 결과를 확인하기에 편리한 패키지 구성을 목표로 개발됩니다. statmanager-kr 개발의 최종 목표는 <b>Python과 Pandas를 잘 알지 못하는 사람</b>도 이용할 수 있는 <b>매우 간편하면서도 유용한 패키지를 만드는 것</b>입니다.  
+Statmanager-kr은 가설을 검증하기 위해 통계 분석이 필요한 연구원, 데이터분석가, 심리학자, 학생 등을 위해 개발되었습니다. statmanager-kr은 사용이 쉽고, 복잡하지 않은 통계 패키지를 목표로 지속적으로 개발됩니다. statmanager-kr 개발의 최종 목표는 <b>Python과 Pandas를 잘 알지 못하는 사람</b>도 이용할 수 있는 <b>매우 간편하면서도 유용한 통계 패키지를 만드는 것</b>입니다.  
 
 Currently, <b><u>KOREAN</u></b> and <b><u>ENGLISH</u></b> are supported.   
 현재 지원하는 언어 세팅은 <b><u>한글</u></b>과 <b><u>영어</u></b>입니다. 
@@ -78,15 +78,13 @@ Using "Jupyter Notebook" is <b>STRONGLY RECOMMENDED</b> (Of course, statmanager-
 
 # Quick Start
 
-
-
 ### Import 
 
 ```Python
 import pandas as pd
 from statmanager import Stat_Manager
 
-df = pd.read_csv('testdf.csv', index_col = 'name')
+df = pd.read_csv('testdf.csv', index_col = 'id')
 sm = Stat_Manager(df, language = 'eng')
 ```
 
@@ -241,6 +239,22 @@ sm.progress(method = 'f_oneway_rm', vars = ['prescore','postscore','fupscore'], 
 ![figure](./doc/output_f_oneway_rm.png)
 
 </details>
+
+<br>
+
+# Related Software
+
+As mentioned earlier, `Statmanager-kr` was developed to provide a user-friendly way to perform statistical analysis methods to test hypotheses, even if the researcher is not familiar with programming languages such as Python. As such, a related software that provides similar user-friendly features is [`Pingouin`](https://pingouin-stats.org/build/html/index.html). 
+
[…]
