### 2024-05-14 ckdckd145/statmanager-kr#13 @ckdckd145
@catstats

Hi! Thank you again for reviewing statmanager-kr.

On point 1, I've fixed the readme file in dev branch. Please check [this commit](https://github.com/ckdckd145/statmanager-kr/commit/2288fc123041a6f8e1d2501ea29c0954e4a45e7b).

And, on point 2, after several tests, it seems that if the version of the `seaborn` library is <b>lower than `0.13.0`</b> (e.g. 0.12.2), a misalignment error occurs. I wasn't aware of this, so thanks for bringing it to my attention and I'll make sure to include it in the official documentation.

Thank you!
