### 2024-05-14 ckdckd145/statmanager-kr#13 @catstats
@ckdckd145
Thank you for your reply.
The README file looks great now, and the misalignment issue in the plot has been resolved after updating Seaborn. I appreciate your thorough testing and prompt updates.
