### 2024-05-15 e90963a
Cleaning Typo Error - "Indenpendent" -> "Independent" in howtouse .csv files - "fgiure" -> "figure" in sentences printed when running .howtouse()
