### 2024-05-15 8db32eb
README UPDATE - Add a recommendation to use the latest versions of dependent libraries and packages
