### 2024-05-15 7750fc2
Update .howtouse() method - Add example code about .progress() - Little revisions in sentences printed when running .howtouse()
