### 2024-05-19 ckdckd145/statmanager-kr#12 @ckdckd145
@behinger

Hi! Thank you again for reviewing statmanager-kr !
Now I'm working on fixing the code, reivising the readme, and documentation based on your valuable feedback.
In particular, I'm working on the automated testing you pointed out (I'm not a developer by nature, so I wasn't aware of these procedures).

By the way, I was wondering if you could tell me directly the code where you got the error you mentioned in `ttest_ind` and `ttest_rel`? I've tried several times, but I'm having a hard time reproducing the error you said.

Please have a look if you have a chance and get in touch with us, thanks very much!
