### 2024-05-19 91c37ef
UPDATE of Levene, Fmax - Now result printed in df format, instead of plain text - Update for pytest
