### 2024-05-19 8c1d7a1
UPDATE of chi2_contingency, fisher - Now result printed in df format, instead of plain text - Update for pytest
