### 2024-05-19 ca84b1f
ADD pytest files - normality test - homoskedasticity test - frequency test
