### 2024-05-29 45c9c7f
ADD pytest files - ttest_ind - ttest_ind_trim - ttest_ind_welch - mannwhitneyu - brunner
