### 2024-05-31 431564d
ADD pytest - f_oneway - kruskal - friedman - rm_anova
