+++ test/test_within_group.py
@@ -2,6 +2,7 @@ import pandas as pd
 from statmanager import Stat_Manager
 import numpy as np
 from scipy import stats
+from statsmodels.stats.anova import AnovaRM
 import ast
 
 import pytest
@@ -69,3 +70,47 @@ def test_wilcoxon():
     assert sm_s == np.round(scipy_s, 3)
     assert sm_pvalue == np.round(scipy_pvalue, 3)
     assert sm_z == np.round(scipy_z, 3)
+
+def test_friedman():
+    '''
+    testing the friedman test (vs. Scipy)
+    '''
+    result_df = sm.progress(method = 'friedman', vars = ['prescore', 'postscore', 'fupscore']).df_results[1]
+    result_object = stats.friedmanchisquare(df['prescore'], df['postscore'], df['fupscore'])
+    
+    scipy_s = result_object.statistic
+    scipy_pvalue = result_object.pvalue
+    
+    sm_s = result_df['correcting for ties'].item()
+    sm_pvalue = result_df['p-value'].item()
+    
+    assert sm_s == np.round(scipy_s, 3)
+    assert sm_pvalue == np.round(scipy_pvalue, 3)
+    
+
+def test_rm_anova():
+    '''
+    testing the One-way Repeated Measures ANOVA (vs. Statsmodels)
+    '''
+    
+    target_columns = ['prescore', 'postscore', 'fupscore']
+    melted_df = df.reset_index().melt(id_vars = 'id', value_vars = target_columns)
+    result_object = AnovaRM(melted_df, depvar = 'value', subject = 'id', within= ['variable']).fit()
+    anova_table = result_object.anova_table
+    
+    result_df = sm.progress(method = 'f_oneway_rm', vars = target_columns).df_results[1]
+    
+    statsmodels_f = anova_table['F Value'].item()
+    statsmodels_numdf = anova_table['Num DF'].item()
+    statsmodels_dendf = anova_table['Den DF'].item()
+    statsmodels_pvalue = anova_table['Pr > F'].item()
+    
+    sm_f = result_df['F Value'].item()
+    sm_numdf = result_df['Num DF'].item()
+    sm_dendf = result_df['Den DF'].item()
+    sm_pvalue = result_df['p-value'].item()
+    
+    assert sm_f == np.round(statsmodels_f, 3)
+    assert sm_numdf == np.round(statsmodels_numdf, 3)
+    assert sm_dendf == np.round(statsmodels_dendf, 3)
+    assert sm_pvalue == np.round(statsmodels_pvalue, 3)
