### 2024-06-05 19d0053
UPDATE of posthoc - now posthoc results will returned as pd.DataFrame instead of simple table in Statsmodels
