+++ statmanager/posthoc_functions.py
@@ -10,20 +10,52 @@ def posthoc_within(df, vars, parametric: bool, posthoc_method):
     posthoc_df = df.reset_index().melt(id_vars=index_col, value_vars=vars)
     mc = MultiComparison(posthoc_df['value'], posthoc_df['variable'])
     
-    if posthoc_method == 'bonf':
-        if parametric:
-            result = mc.allpairtest(stats.ttest_ind, method ='bonf')
-            result_table = result[0]
+    results_for_return = []
+    
+    if parametric:
+        result = mc.allpairtest(stats.ttest_ind, method ='bonf')
+        result_table = result[0]
+        annotation = extract_caption(result_table.as_html())
+        result_table = pd.DataFrame(result_table)
         
-        if not parametric:
-            result = mc.allpairtest(stats.mannwhitneyu, method ='bonf')
-            result_table = result[0]            
+        result_table = result_table.drop(index = 0)
+        result_table.columns = ['group1', 'group2', 'stat', 'p-value', 'corrected p-value', 'reject']
+        result_table['pair no.'] = [n for n in range(1, len(result_table.index) + 1)]
+        result_table = result_table.set_index('pair no.')
+        
+        results_for_return.append(annotation)
+        results_for_return.append(result_table)
+        
+    
+    if not parametric:
+        result = mc.allpairtest(stats.mannwhitneyu, method ='bonf')
+        result_table = result[0] 
+        annotation = extract_caption(result_table.as_html())
+        result_table = pd.DataFrame(result_table)
+        
+        result_table = result_table.drop(index = 0)
+        result_table.columns = ['group1', 'group2', 'stat', 'p-value', 'corrected p-value', 'reject']
+        result_table['pair no.'] = [n for n in range(1, len(result_table.index) + 1)]
+        result_table = result_table.set_index('pair no.')
+        
+        results_for_return.append(annotation)
+        results_for_return.append(result_table)         
     
     elif posthoc_method == 'tukey':
         result = mc.tukeyhsd()
         result_table = result.summary()
         
-    return result_table
+        annotation = extract_caption(result_table.as_html())
+        result_table = pd.DataFrame(result_table)
+        result_table = result_table.drop(index = 0)
+        result_table.columns = ['group1', 'group2', 'mean differences', 'adjusted p-value', 'lower', 'upper', 'reject']
+        result_table['pair no.'] = [n for n in range(1, len(result_table.index) +1 )]
+        result_table = result_table.set_index('pair no.')
+        
+        results_for_return.append(annotation)
+        results_for_return.append(result_table)
+
+    return results_for_return
 
 
 def posthoc_between(df, vars, group_vars, group_names, parametric: bool, posthoc_method):
@@ -38,20 +70,54 @@ def posthoc_between(df, vars, group_vars, group_names, parametric: bool, posthoc
     
     mc = MultiComparison(selected_df[vars], selected_df[group_vars])
     
+    results_for_return = []
+    
     if posthoc_method == 'bonf':
         if parametric:
             result = mc.allpairtest(stats.ttest_ind, method ='bonf')
             result_table = result[0]
+            annotation = extract_caption(result_table.as_html())
+            result_table = pd.DataFrame(result_table)
+            
+            result_table = result_table.drop(index = 0)
+            result_table.columns = ['group1', 'group2', 'stat', 'p-value', 'corrected p-value', 'reject']
+            result_table['pair no.'] = [n for n in range(1, len(result_table.index) + 1)]
+            result_table = result_table.set_index('pair no.')
+            
+            results_for_return.append(annotation)
+            results_for_return.append(result_table)
+            
         
         if not parametric:
             result = mc.allpairtest(stats.mannwhitneyu, method ='bonf')
-            result_table = result[0]            
+            result_table = result[0] 
[…]
