### 2024-06-07 bd1b810
ADD raise errors in between group functions
