### 2024-06-07 ckdckd145/statmanager-kr#12 @ckdckd145
@behinger

Hi!
I just want to say thank you again for your review.

As I said before, I'm still working on fixes to improve my software related to the feedback you mentioned.
You can always find my working in the [dev branch](https://github.com/ckdckd145/statmanager-kr/compare/main...dev) of my repository.
Additionally, the [official documentation](https://www.notion.so/cslee145/Statmanager-kr-Official-Documentation-74a610c12881402d96dc5d1654f97433) and [readme](https://github.com/ckdckd145/statmanager-kr/blob/dev/README.md) are being revised, which I hope you'll check out.
To follow up on the work myself, I'll be utilizing the checkboxes in the feedback you leave in this issue.
Only when everything is complete will I comment again so that you can see each fix and the intent behind it.

I apologize for the significant delay. I'm guessing you've been busy as well and haven't been able to give me the code I need to reproduce the error you comment, so I'm looking to fix the code to raise an error on unintended user errors across the board.
I am also continuing to write the code needed for automated testing with pytest.

Thanks!
