### 2024-06-09 #6642 @teonbrooks
@teonbrooks thanks for the heads up on that.

thanks @catstats and @behinger for your diligent work in reviewing the software and its paper! I really appreciate your hard work and effort.

it looks like there are some outstanding requests @behinger raised in https://github.com/ckdckd145/statmanager-kr/issues/12 that would need to be resolved. I believe that @ckdckd145 is working to resolve these. I think once these remaining issues are resolved, we can do a final pass and make the decision
