### 2024-06-09 ckdckd145/statmanager-kr#12 @teonbrooks
thanks for the review @behinger. could you highlight where you saw the error for the `ttest_ind` and `ttest_rel` when you have a chance?
