### 2024-06-10 ccebe15
UPDATE of posthoc - add code for change types of posthoc dataframe data (statsmodels.iolib.table.Cell -> str | float)
