### 2024-06-10 ckdckd145/statmanager-kr#12 @behinger
hey sorry - I missed the first bump. I think my issue was that running something like this:

`sm.progress(method = 'ttest_rel', vars = 'age', group_vars = 'sex')`

results in an error (now `Keyerror 'a'` instead of `keyerror 's'`) (e.g. replacing the ttest_rel with ttest_ind), because it expects `vars =` and not `group_vars`

I was simply thinking everything is in tidy-data format (one row per observation), but o.c. one can also employ a wider format like here.
