### 2024-06-10 ckdckd145/statmanager-kr#12 @ckdckd145
@behinger

Hi! You don't have to be sorry. :)
Anyway, I finally understand what the not-user-friendly error you're talking about is, and I'll work on fixing the code right away.
Also, I'm working on a fix to prevent errors across the board and provide more user-friendly error messages based on the examples you mentioned.

Also, while the documentation explains why the index column should be set for the user (you can see the related paragraph in the documentation [here](https://www.notion.so/cslee145/Statmanager-kr-Official-Documentation-74a610c12881402d96dc5d1654f97433?pvs=4#8aa12c396b004e508b4b7b51a8398405)), I will explain to you separately why the index column should be set in `Pandas.DataFrame` to use statmanager-kr.

Actually, many of the features of this software owe their origin to `Scipy` and `Statsmodels`, as you may know. The problem is that while `Scipy` is working in a way that is compatible with what I call a wide-frame dataframe (where the information obtained from each subject is arranged horizontally), `Statsmodels` is not. As far as I know, in order to use the functionality of `Statsmodels`, in many cases it is necessary to change the `Pandas.DataFrame` from 'wide-range' to 'tidy-data', which is usually referred to in database languages like `SQL`. In Statmanager-kr, this is done via methods like `melt()` in `Pandas`, but the problem is that `pd.melt()` always requires an index parameter, which means that in order to use the functionality of `Statsmodels` (because the dataframe must be reframed via `melt()` to compatible with `Statsmodels`), I need to obtain information about the column that serves as the index.

I'm not sure if I've explained this correctly, but I hope this helps address some of the comments you've made.
I'll keep you posted on my work

Thank you!
