### 2024-06-10 705822f
ADD codes for raising errors in between, within group functions
