### 2024-06-10 ckdckd145/statmanager-kr#12 @behinger
ah, I must have missed the explanation, very reasonable, thanks for repeating for me
