### 2024-06-11 d30f97f
ADD codes for error raising errors in some functions
