### 2024-06-12 7fe41da
ADD codes for error raising in ANCOVA functions
