### 2024-06-17 410f5e8
Revision : posthoc functions and ancova functions - fix little bugs in posthoc tables or posthoc functions
