### 2024-06-17 20d06a1
ADD pytest - bonf, tukey hsd in kruskal-wallis - bonf, tukey hsd in oneway-ancova - bonf, tukey hsd in Oneway RM anova - bonf, tukey hsd in friedman
