### 2024-06-19 d2dc2f4
ADD pytests - posthoc in rm_ancova - posthoc in f_nway
