### 2024-06-20 d259e2a
ADD pytests - posthoc in f_nway_rm
