+++ test/test_posthoc.py
@@ -15,8 +15,6 @@ df = pd.read_csv(r"./testdata/testdf.csv", index_col = 'id')
 sm = Stat_Manager(df)
 
 
-# f_nway_rm
-
 
 def test_bonf_in_f_oneway():
     '''
@@ -674,5 +672,122 @@ def test_tukey_f_nway():
         
         for n in range(len(statmanager)):
             assert statmanager[n] == statsmodels[n]
-            
+
+def test_bonf_f_nway_rm():
+    '''
+    testing the bonf in f_nway_rm (vs. Statsmodels/Scipy)
+    '''
     
+    dvs = ['prescore', 'postscore', 'fupscore']
+    group_vars = 'condition'
+    
+    result_dfs = sm.progress(method = 'f_nway_rm', vars = dvs, group_vars = 'condition', posthoc = True).df_results[-3:]
+
+    
+    # main-effect-posthoc (repeated-measures)
+    
+    index_col = df.index.name
+    
+    posthoc_df = df.reset_index().melt(id_vars = index_col, value_vars = dvs)
+    mc = MultiComparison(posthoc_df['value'], posthoc_df['variable'])
+    
+    result = mc.allpairtest(stats.ttest_ind, method = 'bonf')
+    result_table = pd.DataFrame(result[0])
+    original_columns = ['group1', 'group2', 'stat', 'pval', 'pval_corr', 'reject']
+    result_table = result_table.drop(index = 0)
+    result_table.columns = original_columns
+    
+    for index in result_table.index:
+        for column in result_table.columns:
+            result_table.loc[index, column] = result_table.loc[index, column].data
+    
+    result_df = result_dfs[1]
+    
+    for n in range(len(result_table.index)):
+        statmanager = result_df.iloc[n].to_list()
+        statsmodels = result_table.iloc[n].to_list()
+        
+        for n in range(len(statmanager)):
+            assert statmanager[n] == statsmodels[n]
+            
+    # main-effect-posthoc (group variable)
+    
+    posthoc_df = df.reset_index().melt(id_vars = index_col, value_vars = dvs, var_name = 'time').set_index(index_col)
+    merged_df = df.drop(columns = dvs).merge(posthoc_df, how = 'outer', on = index_col)
+    new_dv = 'value'
+
+    mc = MultiComparison(merged_df[new_dv], merged_df['condition'])
+    result = mc.allpairtest(stats.ttest_ind, method = 'bonf')
+    result_table = pd.DataFrame(result[0])
+    original_columns = ['group1', 'group2', 'stat', 'pval', 'pval_corr', 'reject']
+    result_table = result_table.drop(index = 0)
+    result_table.columns = original_columns
+
+    for index in result_table.index:
+        for column in result_table.columns:
+            result_table.loc[index, column] = result_table.loc[index, column].data
+
+
+    result_df = result_dfs[0]
+        
+    for n in range(len(result_table.index)):
+        statmanager = result_df.iloc[n].to_list()
+        statsmodels = result_table.iloc[n].to_list()
+        
+        for n in range(len(statmanager)):
+            assert statmanager[n] == statsmodels[n]
+    
+    
+    # interaction
+    index_col = df.index.name
+    group_vars = ['condition', 'time']
+
+
+    posthoc_df = df.reset_index().melt(id_vars = index_col, value_vars = dvs, var_name = 'time').set_index(index_col)
+    merged_df = df.drop(columns = dvs).merge(posthoc_df, how = 'outer', on = index_col)
+
+    interactions = []
+    new_df = merged_df.copy()
+
+    for n in range(2, len(group_vars) + 1):
+        for combo in combinations(group_vars, n):
+            interaction_name = "interaction_" + "_".join(combo)
+            
[…]
