### 2024-06-26 ckdckd145/statmanager-kr#12 @ckdckd145
@behinger

Hi!

Letting you know how your work is progressing.
I've almost finished creating the Python files for the automatic tests. The only thing left to do is to update and deploy statmanager-kr to ensure this process goes smoothly, and set up the automatic tests to run on the Github repository. Anyway, my plan is to have the process automatically check for compatible Python versions as well. I'm struggling a bit with this, as I've never done this on Github before. The python files I wrote for the automatic tests can be found in the [dev branch](https://github.com/ckdckd145/statmanager-kr/tree/dev/test).

By the way, I will add things related to the "Community guidelines" in readme file and documentation.

Thank you!
