### 2024-07-04 ba89497
forget to provide the posthoc_method parameter in pytest file for testing the posthoc methods
