### 2024-07-05 edc9aaf
add pytest files - effectsize testing
