### 2024-07-05 e8af545
badge refreshing + setup.py / init change for releasing
