### 2024-07-05 640c9f4
Create automatic_test.yml
add automatic testing actions in Github
