### 2024-07-05 0dc7e4c
Add available OS and Python versions in readme.md
