+++ README.md
@@ -1,8 +1,19 @@
  
-[![PyPI version](https://badge.fury.io/py/statmanager-kr.svg)](https://badge.fury.io/py/statmanager-kr)
+[![PyPI version](https://badge.fury.io/py/statmanager-kr.svg?kill=cache)](https://badge.fury.io/py/statmanager-kr)
 [![license](https://img.shields.io/badge/license-MIT-green.svg)](https://github.com/ckdckd145/statmanager-kr/blob/main/LICENSE)
 ![logo](./doc/logo.png)
 
+
+### Available Operating Systems
+`Windows 10`   
+`Windows 11`   
+`Mac`   
+`Linux`  
+
+### Availabe Python Versions
+`Python 3.10`   
+`Python 3.11`   
+`Python 3.12`
 
 # Statmanager-kr
 ### Open-source Statistical Package for Python based on the Pandas.    
