### 2024-07-05 956ad10
Revert "Add available OS and Python versions in readme.md"
This reverts commit 0dc7e4c531f7e94886793d25e604d4639c8ce28f.
