### 2024-07-05 671cdb7
add os/python versions info in readme
