### 2024-07-05 ckdckd145/statmanager-kr#12 @ckdckd145
@behinger

Hi!

I deeply appreciate your patience and wait.
Finally, I finalized the code for automatic testing, created a github action to run it, and saw all the tests pass. (I'm very excited)
Also, I add the available python versions validated via testing in readme.md.

As if that wasn't enough, I also added a contribute guideline to the documentation. I really hope these fixes will help you finalize your review.

Have a great day!
Thank you.

ref @teonbrooks
