### 2024-07-08 #6642 @behinger
With https://github.com/ckdckd145/statmanager-kr/issues/12 I think my concerns were nicely addressed
