### 2024-07-08 ckdckd145/statmanager-kr#12 @behinger
nice! I answered the JOSS issue. My points were addressed.

Maybe a small note: The "how to contribute" is one step dettached from the github readme in the notion file, maybe worth linking to that section from the readme immediately.
