### 2024-07-11 #6642 @ckdckd145
@teonbrooks

Hi! ✋🏼

I apologize for the long delay, but it looks like the review has been completed, can you please tell me what to do now, I'm new to the JOSS process so I'm not sure.

Thank you very much 😄
