### 2024-07-11 #6642 @teonbrooks
@ckdckd145, no worries, it has been great watching how both the package and paper have improved over the review cycle.

@catstats, anything you would like to add or are you happy where things have landed?

@behinger, I take it that you are satisfied with the paper and package.
