### 2024-07-15 eb3f97c
Direct link about contribution guidelines in readme.md | Little revisions in article
