+++ article/paper.md
@@ -19,17 +19,17 @@ bibliography: paper.bib
 
 # Summary
 
-Python is one of the most popular and easiest programming languages. Despite the fact that manay researchers use Python for data preprocessing and statistical analysis, there are few statistical packages that have inherited the simple, easy-to-use nature of Python.  Consequently, people who possess statistical knowledge but lack familiarity with programming languages continue to rely on other costly software. Because researchers who are not familiar with programming may not know how to use different methods and adjust parameters effectively. 
+Python is one of the most popular and easy-to-use programming languages. Despite the fact that many researchers use Python for data preprocessing and statistical analysis, there are few statistical packages that have inherited the simple and easy-to-use nature of Python. As a result, people with statistical skills but no familiarity with programming languages continue to rely on other expensive software. This is because researchers who are not familiar with programming may not know how to use different methods and adjust parameters effectively. 
 
-The `Statmanager-kr` was designed to provide easy-to-use statistical functions for people with little knowledge of programming languages. Because many people are already familiar with data in table format, such as that in Microsoft Excel, `Statmanager-kr` was designed to be compatible with `Pandas.DataFrame`. In addition, the `Statmanager-kr` was designed so that the analysis is performed using only one method utilizing as few parameters as possible.
+The `Statmanager-kr` has been designed to provide easy-to-use statistical functions especially for people with little knowledge of programming languages. The `Statmanager-kr` was designed to be compatible with `Pandas.DataFrame`. In addition, the `Statmanager-kr` was designed so that the analysis is performed using only one method with as few parameters as possible.
 
-Additionally, `Statmanager-kr` relies on `Scipy` and `Statsmodels` for valid analysis. The `Statmanager-kr` provides methods related to testing the normality and homoscedasticity assumptions, comparing between-group and within-group differences, conducting regression analysis, and data visualization.
+Additionally, `Statmanager-kr` relies on `Scipy` and `Statsmodels` for valid analysis. The `Statmanager-kr` provides methods for testing the normality and homoscedasticity assumptions, comparing between-group and within-group differences, performing regression analysis, and data visualization.
 
 # Statement of need
 
-The `statmanager-kr` is a statistical package for Python in `Pandas`. This package provides methods commonly used for null hypothesis significance testing (NHST), which is of interest to researchers in various fields [@Moon2020]. It is also possible to check whether the normality or equivariance assumption is met, using the Shapiro-Wilk, Levene, or Fmax tests. 
+The `statmanager-kr` is a statistical package for Python in `Pandas`. This package provides methods commonly used for null hypothesis significance testing (NHST), which is of interest to researchers in various fields [@Moon2020]. It is also possible to test for normality or equivariance using the Shapiro-Wilk, Levene, or Fmax tests. 
 
-Most of the statistical software available today is difficult to use, as demonstrated by previous study reporting that one of the challenges college students face in statistics courses was "using software" [@Murtonen:2003]. Although there are basic statistical libraries in Python, such as Scipy [@seabold:2010] and Statsmodels [@Virtanen:2020], they are quite complex. While some studies involve complex and detailed statistical modeling and analysis, there are also many studies that require only a few hypothesis tests. Therefore, the development of an easy-to-use statistical package would be of great benefit to these researchers. 
[…]
