### 2024-07-15 #6642 @ckdckd145
@teonbrooks

Hi!
I made little revisions to the article to remove typos and misrepresentations. 😄
