### 2024-07-26 #6642 @teonbrooks
@crvernon I noticed some formatting issues in the pdf rendering of the manuscript, specifically, I see some overflow of text in some columns. are there ways to have the typesetting adjusted? Is that done at this stage?
