### 2024-07-26 #6642 @crvernon
@teonbrooks this is the point in the review that you do your pass and once you are satisfied you comment the following in the thread which notifies me that all is ready for final approval.  Here are the things I normally check when I edit that you can do now:

You still need to ensure these things are accounted for.  You should generate your own checklist using the command found here:  https://joss.readthedocs.io/en/latest/editorial_bot.html#post-review-checklist

This also tells you how to set the version and DOI for the archive.  Feel free to follow up on Slack if you have any questions.
