### 2024-08-09 166f68b
Update paper.md
I added some copy-editing corrections to the body of the article.
