+++ article/paper.md
@@ -2,7 +2,7 @@
 title: 'Statmanager-kr: A User-friendly Statistical Package for Python in Pandas'
 tags:
   - Python
-  - statistic analysis
+  - statistical analysis
   - social science
   - null-hypothesis
   - user-friendly
@@ -21,17 +21,17 @@ bibliography: paper.bib
 
 Python is one of the most popular and easy-to-use programming languages. Despite the fact that many researchers use Python for data preprocessing and statistical analysis, there are few statistical packages that have inherited the simple and easy-to-use nature of Python. As a result, people with statistical skills but no familiarity with programming languages continue to rely on other expensive software. This is because researchers who are not familiar with programming may not know how to use different methods and adjust parameters effectively. 
 
-The `Statmanager-kr` has been designed to provide easy-to-use statistical functions especially for people with little knowledge of programming languages. The `Statmanager-kr` was designed to be compatible with `Pandas.DataFrame`. In addition, the `Statmanager-kr` was designed so that the analysis is performed using only one method with as few parameters as possible.
+`Statmanager-kr` has been designed to provide easy-to-use statistical functions especially for people with little knowledge of programming languages. `Statmanager-kr` was designed to be compatible with `Pandas.DataFrame`. In addition, the `Statmanager-kr` was designed so that the analysis is performed using only one method with as few parameters as possible.
 
-Additionally, `Statmanager-kr` relies on `Scipy` and `Statsmodels` for valid analysis. The `Statmanager-kr` provides methods for testing the normality and homoscedasticity assumptions, comparing between-group and within-group differences, performing regression analysis, and data visualization.
+Additionally, `Statmanager-kr` relies on `Scipy` and `Statsmodels` for valid analysis. `Statmanager-kr` provides methods for testing the normality and homoscedasticity assumptions, comparing between-group and within-group differences, performing regression analysis, and data visualization.
 
 # Statement of need
 
-The `statmanager-kr` is a statistical package for Python in `Pandas`. This package provides methods commonly used for null hypothesis significance testing (NHST), which is of interest to researchers in various fields [@Moon2020]. It is also possible to test for normality or equivariance using the Shapiro-Wilk, Levene, or Fmax tests. 
+`Statmanager-kr` is a statistical package for Python in `Pandas`. This package provides methods commonly used for null hypothesis significance testing (NHST), which is of interest to researchers in various fields [@Moon2020]. It is also possible to test for normality or equivariance using the Shapiro-Wilk, Levene, or Fmax tests. 
 
 Most of the statistical software available today is difficult to use, as a previous study reported that one of the challenges students face in statistics courses was "using software" [@Murtonen:2003]. Although there are basic statistical libraries in Python, such as Scipy [@seabold:2010] and Statsmodels [@Virtanen:2020], they are quite complex. While some studies require complex and detailed statistical modeling and analysis, there are also many studies that require only a few hypothesis tests. Therefore, the development of an easy-to-use statistical package would be of great benefit to these researchers. 
 
-To achieve this, the `statmanager-kr` has been designed to run analyses with only three lines of code: 1. read data as a `Pandas.DataFrame`, 2. create a `Stat_Manager` object, 3. execute the `.progress()` method. Therefore, users can use the `Statmanager-kr` as long as they know the Pandas methods to read the data, such as `.read_csv()` or `.read_excel()`. It also includes functions to visualize the results depending on the analysis method.
[…]
