### 2024-08-28 #6642 @ckdckd145
@teonbrooks

I hope you're having some good days.
Is it okay if I revise the abstract of an article now?
I don't know the current process, so I'm not sure if it's okay.

Thanks!
