### 2024-08-30 #6642 @teonbrooks
@ckdckd145, would you like to make a new version of your package to capture the manuscript changing? I would do it after you finish the abstract edits.
