### 2024-09-01 #6642 @ckdckd145
@teonbrooks

Hi!
Yes, exactly. I will make a ner version for abstract revision.
The content of abstract will not be changed, just I want to improve the quality of the sentences :)

I will notify after ther release a new version.
Maybe this probably won't take long.
Thanks !
