### 2024-09-02 b6324b8
update in readme and article.md
