### 2024-09-02 #6642 @ckdckd145
@teonbrooks

Hi! I make a revision in my github repository.
Also, I generate the new pdf.

By the way, I find the erorrs in importing `statsmodels` in macOS.
I did a search and found that others are also experiencing this error on macOS. [[Link]](https://github.com/statsmodels/statsmodels/issues/9333)

I'll keep an eye on it and make sure there are no issues.

Anyway, the abstract modification is done.
If you're able, I'd appreciate it if you could go through the following process. :)

Thanks!
