### 2024-09-09 #6642 @teonbrooks
@crvernon I noticed on page 4, there is a column formatting error where the y in `chi2_contingency` is colliding with `Chi-squared Test` in the second column. is that a formatting issue we can fix?
