### 2024-09-09 #6642 @crvernon
Hi @teonbrooks - I would think the `Analysis` column would overflow but it does not.  I can ask around about this and run a few tests myself.  However, I think the "Features" section of this paper can be removed all together because all of this info is in the documentation for the package itself.  I would recommend removing it to make the paper cleaner and closer to our target 1000 word limit.  If the authors wish to do so, they can mention the docs in the paper and provide a link to them.

Let me know what you think.

Thanks!
