### 2024-09-10 #6642 @crvernon
Yes, I think they could be removed.  But that is your and the author's call.  I believe that the fundamental parts of the publication are there without those sections and that they best serve the reader as part of the software documentation.  This is where JOSS differs a bit from other journals.  But again, I will support what you decide in this case.
