### 2024-09-17 #6642 @teonbrooks
@ckdckd145, how would you feel about replacing the sections with `Features` in the paper with the links to the sections in the documentation:
e.g. `User Friendly features` points to https://cslee145.notion.site/Statmanager-kr-Official-Documentation-74a610c12881402d96dc5d1654f97433 instead of the text currently there and https://cslee145.notion.site/Statmanager-kr-dded43262f784c70a37fddb11ec7c9d1

This would make the paper more concise and it will adapt with any feature changes.

After thinking about it a bit more, I'm more inclined to leave it in the documentation and link to it there than it being in the paper but I wanted to get your thoughts
