### 2024-09-24 c969be5
update for the article revision (1.8.1.14.)
