+++ article/paper.md
@@ -42,115 +42,9 @@ In conclusion, depending on the researcher's programming experience and the purp
 
 # Features
 
-`Statmanager-kr` was designed to be compatible with the wide range form of `Pandas.DataFrame`.  
-
-## User-friendly Features
-
-### Setting the Language
-
-It is possible to change the language by adjusting the `language` parameter when creating an object of the Stat_Manager class. The supported languages are Korean ("kor") and English ("eng"), and the default is Korean.
-
-```python
-sm = Stat_Manager(df, language = 'eng')
-```
-
-### Other Methods
-
-Users can search for a specific usage by calling the `.howtouse()` method. It can also change the language with `.set_language()`, or change the dataframe by running `.change_dataframe()`.  
+Learn more about the features in the *[official documentation](https://cslee145.notion.site/60cbfcbc90614fe990e02ab8340630cc?v=4991650ae5ce4427a215d1043802f5c0)*. The instructions for running each analysis are also described in *[the manual section of the official documentation](https://cslee145.notion.site/Statmanager-kr-Official-Documentation-74a610c12881402d96dc5d1654f97433#be93db7f4159419fa73eb324d6567793)*. *[Korean versions of the official documenation](https://cslee145.notion.site/fd776d4f9a4f4c9db2cf1bbe60726971?v=3b2b237555fc4cd3a41a8da337d80c01)* are also available.
 
 
-## Statistical Test
-
-The implementation of analysis in `Statmanager-kr` can be summarized as follows.
-
-| Objective                             | Analysis                                                                                                                                                |
-| ------------------------------------- | ------------------------------------------------------------------------------------------------------------------------------------------------------- |
-| Check the normality assumption        | Kolmogorov-Smirnov Test, Shapiro-Wilks Test, Z-Skewness & Z-Kurtosis Test                                                                              |
-| Check the homoskedasticity assumption | Levene's Test, F<sub>max</sub> Test                                                                                                                                 |
-| Frequency analysis                    | Chi-Squared Test, Fisher’s Exact Test                                                                                                                   |
-| Check the reliability of the scale    | Calculating Cronbach’s Alpha                                                                                                                            |
-| Correlation analysis                  | Pearson’s r, Spearman’s rho, Kendall’s tau                                                                                                              |
-| Comparison between groups             | Independent Samples T-test, Yuen’s T-test, Welch’s T-test, Mann-whitney U test, Brunner-Munzel Test, One-way ANOVA, Kruskal Wallis Test, One-way ANCOVA |
-| Comparison within group               | Dependent Samples T-test, Wilcoxon-Signed Rank Test, One-way Repeated Measures ANOVA, Friedman Test, Repeated Measures ANCOVA,                          |
-| Comparison by multiple ways           | N-way ANOVA, N-way Mixed Repeated Measures ANOVA                                                                                                        |
-| Regression analysis                   | Linear Regression, Logistic Regression                                                                                                                  |
-| etc                                   | Bootstrapping percentile method                                                                                                                         |
-
[…]
