### 2024-09-24 #6642 @ckdckd145
@teonbrooks
I'm excited that it's almost done 😄
I made the revisions in article, also I made the new version tage ([1.8.1.14](https://github.com/ckdckd145/statmanager-kr/releases/tag/1.8.1.14)).

Thanks! 👍🏼
