### 2024-09-26 #6642 @ckdckd145
@teonbrooks I made the DOI with zenodo !
[![DOI](https://zenodo.org/badge/725431811.svg)](https://zenodo.org/doi/10.5281/zenodo.13840988)

Thanks 😄
