### 2024-09-26 #6642 @ckdckd145
@teonbrooks

Thank you so much for your help! I'm very excited to publish my article in JOSS 😄
Also, special thanks to @behinger @catstats for reviewing my software 👍🏼
