### 2024-10-07 #6642 @crvernon
:wave: @ckdckd145 - I just need you to address the following before I move forward with accepting this for publication:

**In your archive:**
- [ ] The title and authors (including the author order) in your Zenodo archive should match those represented in your paper exactly.  Please edit the Zenodo archive's metadata to match these fields to what is in your paper.
- [ ] Also, the license specified in your Zenodo archive should match what is used in your code repository.

**In your paper:**
- [ ] LINE 12: "Scipy" should be "SciPy"
- [ ] LINE 12: "Statsmodels" should be "statsmodels"; I believe the capitalized the "S" in the title of the journal article because it was the start of a sentence; however, the package has it all lower case.
- [ ] LINE 23:  "Scipy" should be "SciPy"
- [ ] LINE 24: "Statsmodels" should be "statsmodels"
- [ ] LINES 23 and 24:  The citations for SciPy and statsmodels are switched up.
- [ ] LINE 29:  Use ")" instead of "." after your numbered list.  For example, "...1) read data..."
- [ ] LINE 30:  No need for the dot in front of the method name or the parenthesis after; code formatting is ok.  So it should read:  "3) execute the `progress` method..."
- [ ] LINE 31:  No need for the dot in front of the method names for Pandas functions or the parenthesis after; code formatting is ok.   Also, I believe these are functions and not methods as a part of a class.
- [ ] LINE 46:  I would change "datasets" here to "data types" because you are directly reference the Pandas `DataFrame` data type here.
- [ ] LINE 47-49:  Missing content after "post-hoc".  Are you referring to post hoc processing, e.g., post processing?  These few sentences read a bit off.  How about the following as a recommended fix at the start of the paragraph:
>`Statmanager-kr` also is able to perform additional post processing directly by setting `posthoc` to `True` in the `progress` method.  Results can then be visualized by calling the `figure` method from the output variable or instance of `progress`. Although...
- [ ] Throughout, no need to place the dot before method names and the parenthesis after when you are discussing them as a method or function in the sentence and using code formatting on their names.  I'll not mention these from now on, so please just check the document for these.
- [ ] LINE 72: the "p" in "python" should be capitalized.  This can be done in the bib file by using curly brackets around the characters that you wish to maintain formatting of.
- [ ] LINE 78: the "p" in "python" should be capitalized.

Let me know when you have finished these and I'll take another look.  Thanks!
