+++ article/paper.bib
@@ -28,7 +28,7 @@
   doi={10.1038/s41592-019-0686-2}
 }
 @article{seabold:2010,
-  title={Statsmodels: Econometric and statistical modeling with python},
+  title={Statsmodels: Econometric and statistical modeling with Python},
   author={Seabold, Skipper and Perktold, Josef},
   journal={Proceedings of the 9th Python in Science Conference},
   volume={57},
