### 2024-10-08 #6642 @ckdckd145
@crvernon

Hi! ✋🏼 I appreciate your feedback.
I just make revisions in my articles.
Lines related to the feedback about "post-hoc" also revised to make it clear. Actually it was about post-hoc analysis in scientific statistics. Let me give you the [link](https://en.wikipedia.org/wiki/Post_hoc_analysis) about post-hoc analysis.

And... The DOI changed when I issued a new release. 😢
([10.5281/zenodo.13901218](https://doi.org/10.5281/zenodo.13901218))
I think this related to my unfamiliarity with zenodo. Sorry about that 😓

Thank you!

ref @teonbrooks
